+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ Next Release (TBD)
 * Fix issue where requestParameters were not being mapped
   correctly resulting in invalid generated javascript SDKs
   (`#498 <https://github.com/aws/chalice/issues/498>`__)
+* Fix issue where ``api_gateway_stage`` was being
+  ignored when set in the ``config.json`` file
+  (`#495 <https://github.com/aws/chalice/issues/495>`__)
 
 
 1.0.1